refactor: audio_bridge/handles.rs -> handle.rs
